requirements.txt
streamlit
pypdf
openai
python-dotenv
supabase
stripe
